bug fixing

+++ database/migrations/2021_06_06_201231_create_user_infos_table.php
@@ -23,7 +23,7 @@ class CreateUserInfosTable extends Migration
             $table->string('mobile');
             $table->string('reason');
             $table->string('team');
-            $table->integer('university_id');
+            $table->bigInteger('university_id');
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')
                 ->references('id')
